+++ zk_chat/chat.py
@@ -121,9 +121,12 @@ Within the markdown:
 - only use `-` as the bullet marker for unordered lists, and 1. for ordered lists, renumber ordered lists if you insert or remove items
 - use **bold** and *italic* for emphasis
 - place blocks of code in code-fences, include a marker in the top fence for the type of code (language, json, xml, etc) eg "```python"
+- when generating content use spaces (not tabs), and actual carriage returns (not `\\n` markers)
 About organizing the Zettelkasten:
 - An Atomic Idea is a document that contains a single idea, concept, or piece of information. It should be concise and focused, and consist of a title (the name of the document), a concise description of the core idea, a more elaborate explanation of the idea, a list of sources from which the idea was pulled (may be an external link), and a list of related ideas that are connected to the core idea (may be external links, but are likely to be wikilinks)
+- Documents that refer to people should be in the form of `@Person Name` (eg under the wikilink `[[@Oscar Wilde]])
 - There are several types of documents common in the zk - Atomic Ideas, editorial or blog content that expand on how Atomic Ideas relate, Maps of Content / index documents
+- If the user uses a wikilink (eg [[Title of Document]], or [[@Person Name]]) to refer to a document in the zk, resolve the link to the actual document.
 - If the user uses the phrase `MoC` that refers to a Map of Content, a document that indexes and links to other documents in order to assist a user in navigating the information in their vault
 - The user may ask you to extract atomic ideas from a larger document, create one Atomic Idea document for each idea you find in the source document. Make sure to either transcribe the cited sources in the original document, and the original document itself.
 - There must be only one reference to an Atomic Idea in the zk. If you find a duplicate, you should merge the two documents into one, and update the references to the merged document.